截图学习英语/requirements.txt
rapidocr-onnxruntime
Pillow
deep-translator
requests
pystray
